+++ apps/platform/pkg/meta/bot.go
@@ -67,8 +67,8 @@ type Bot struct {
 	Workspace     string            `yaml:"-" uesio:"studio.workspaceid"`
 	CreatedBy     *User             `yaml:"-" uesio:"studio.createdby"`
 	UpdatedBy     *User             `yaml:"-" uesio:"studio.updatedby"`
-	UpdatedAt     int64             `yaml:"-" uesio:"studio.updatedat"`
-	CreatedAt     int64             `yaml:"-" uesio:"studio.createdat"`
+	UpdatedAt     float64           `yaml:"-" uesio:"studio.updatedat"`
+	CreatedAt     float64           `yaml:"-" uesio:"studio.createdat"`
 	itemMeta      *ItemMeta         `yaml:"-" uesio:"-"`
 }
