+++ apps/platform/pkg/meta/bundle.go
@@ -53,8 +53,8 @@ type Bundle struct {
 	itemMeta    *ItemMeta `yaml:"-" uesio:"-"`
 	CreatedBy   *User     `uesio:"uesio.createdby"`
 	UpdatedBy   *User     `uesio:"uesio.updatedby"`
-	UpdatedAt   int64     `uesio:"uesio.updatedat"`
-	CreatedAt   int64     `uesio:"uesio.createdat"`
+	UpdatedAt   float64   `uesio:"uesio.updatedat"`
+	CreatedAt   float64   `uesio:"uesio.createdat"`
 }
 
 func (b *Bundle) GetVersionString() string {
