+++ apps/platform/pkg/meta/field.go
@@ -51,8 +51,8 @@ type Field struct {
 	itemMeta             *ItemMeta  `yaml:"-" uesio:"-"`
 	CreatedBy            *User      `yaml:"-" uesio:"studio.createdby"`
 	UpdatedBy            *User      `yaml:"-" uesio:"studio.updatedby"`
-	UpdatedAt            int64      `yaml:"-" uesio:"studio.updatedat"`
-	CreatedAt            int64      `yaml:"-" uesio:"studio.createdat"`
+	UpdatedAt            float64    `yaml:"-" uesio:"studio.updatedat"`
+	CreatedAt            float64    `yaml:"-" uesio:"studio.createdat"`
 	SubFields            []SubField `yaml:"subfields" uesio:"studio.subfields"`
 }
