+++ apps/platform/pkg/meta/permissionset.go
@@ -32,8 +32,8 @@ type PermissionSet struct {
 	itemMeta       *ItemMeta       `yaml:"-" uesio:"-"`
 	CreatedBy      *User           `yaml:"-" uesio:"studio.createdby"`
 	UpdatedBy      *User           `yaml:"-" uesio:"studio.updatedby"`
-	UpdatedAt      int64           `yaml:"-" uesio:"studio.updatedat"`
-	CreatedAt      int64           `yaml:"-" uesio:"studio.createdat"`
+	UpdatedAt      float64         `yaml:"-" uesio:"studio.updatedat"`
+	CreatedAt      float64         `yaml:"-" uesio:"studio.createdat"`
 }
 
 // GetCollectionName function
