+++ apps/platform/pkg/meta/secret.go
@@ -13,8 +13,8 @@ type Secret struct {
 	itemMeta  *ItemMeta `yaml:"-" uesio:"-"`
 	CreatedBy *User     `yaml:"-" uesio:"studio.createdby"`
 	UpdatedBy *User     `yaml:"-" uesio:"studio.updatedby"`
-	UpdatedAt int64     `yaml:"-" uesio:"studio.updatedat"`
-	CreatedAt int64     `yaml:"-" uesio:"studio.createdat"`
+	UpdatedAt float64   `yaml:"-" uesio:"studio.updatedat"`
+	CreatedAt float64   `yaml:"-" uesio:"studio.createdat"`
 }
 
 // NewSecret function
