+++ apps/platform/pkg/meta/selectlist.go
@@ -15,8 +15,8 @@ type SelectList struct {
 	Workspace string             `yaml:"-" uesio:"studio.workspaceid"`
 	CreatedBy *User              `yaml:"-" uesio:"studio.createdby"`
 	UpdatedBy *User              `yaml:"-" uesio:"studio.updatedby"`
-	UpdatedAt int64              `yaml:"-" uesio:"studio.updatedat"`
-	CreatedAt int64              `yaml:"-" uesio:"studio.createdat"`
+	UpdatedAt float64            `yaml:"-" uesio:"studio.updatedat"`
+	CreatedAt float64            `yaml:"-" uesio:"studio.createdat"`
 	itemMeta  *ItemMeta          `yaml:"-" uesio:"-"`
 }
